full functional be

+++ manager/persistence.js
@@ -2,17 +2,23 @@ import fs from 'fs'
 
 export default class Persistence {
 
-    static checkWithdrawal(balance) {
-        // console.log('hello')
+    static checkWithdrawal(balance, sum) {
+        return balance > sum
     }
 
     static makeWithdrawal(id, sum) {
-        // console.log("hello")
+        const balance = this.getBalanceFromAccountID(id)
+        const finBal = balance - sum
+        fs.writeFile('./db/' + id + '.txt', finBal.toString(), err => {
+            if (err) {
+                console.error(err);
+            }
+            // file written successfully
+        });
     }
 
 
     static makeDeposit(id, sum) {
-        // console.log("hello")
         const balance = this.getBalanceFromAccountID(id)
         const finBal = balance + sum
 
@@ -27,19 +33,16 @@ export default class Persistence {
     static getBalanceFromAccountID(id) {
         // read from the db directory the files and get the result
         let sum = 10000
-        try {
+        if (fs.existsSync('./db/' + id + '.txt')) {
             sum = fs.readFileSync('./db/' + id + '.txt',
-                { encoding: 'utf8', flag: 'r' });
-        } catch (e) {
-            console.log(e)
+                {encoding: 'utf8', flag: 'r'});
+        } else {
             fs.writeFileSync('./db/' + id + '.txt', '10000')
         }
 
-        if (typeof sum === 'string' || sum instanceof String){
+        if (typeof sum === 'string' || sum instanceof String) {
             sum = parseInt(sum)
         }
-        console.log(sum)
-        console.log(typeof sum)
 
         return sum
     }
